Replace TrailingComment* with _? except after statement delimiters.
TrailingComment includes // comments and those can't appear in the middle of a single line statement
(it would comment out the rest of the line). So now we only check for inline comments in most places
except statement delimiters which should have the rest of line trailing comments attached to them so
they don't get attached to enclosing blocks.